init/resize values accept int instead of size_t to handle negative graph size

+++ data.h
@@ -12,8 +12,8 @@ typedef struct values {
   int i;
 } values_t;
 
-void init_values(values_t *values, size_t history_size);
-void resize_values(values_t *values, size_t history_size);
+void init_values(values_t *values, int history_size);
+void resize_values(values_t *values, int history_size);
 void add_value(values_t *values, float v);
 float previous_value(values_t *values);
 void dealloc_values(values_t *values);
